:recycle: Improve the user use_cases

+++ blog_app/use_cases/user/delete_user.py
@@ -4,7 +4,9 @@ import os
 import shutil
 
 def delete_user(user: User) -> bool:
-    shutil.rmtree(user.get_folder_path())
+    if os.path.exists(user.get_folder_path()):
+        print("Deleting user folder")
+        shutil.rmtree(user.get_folder_path())
     db.session.delete(user)
     db.session.commit()
     return True
